+++ deepparse/cli/parser_arguments_adder.py
@@ -107,8 +107,8 @@ def add_batch_size_arg(parser: ArgumentParser) -> None:
 
 def add_path_to_retrained_model_arg(parser: ArgumentParser) -> None:
     parser.add_argument(
-        "--path_to_model_weights",
-        help=wrap("A path to a retrained model to use for testing."),
+        "--path_to_retrained_model",
+        help=wrap("A path to a retrained model to use. It can be an S3-URI."),
         type=str,
         default=None,
     )