+++ tests/test_weights_tools.py
@@ -1,16 +1,18 @@
 # pylint: disable=too-many-public-methods
+import os
 
 # Pylint error for TemporaryDirectory ask for with statement
 # pylint: disable=consider-using-with
 
 
 import unittest
-from unittest import TestCase
+from unittest import TestCase, skipIf
 from unittest.mock import MagicMock, patch, call
 
-from cloudpathlib import S3Path
+from cloudpathlib import S3Path, CloudPath
 
 from deepparse import handle_weights_upload
+from deepparse.parser import AddressParser, FormattedParsedAddress
 
 
 class WeightsToolsTests(TestCase):
@@ -62,6 +64,23 @@ class WeightsToolsTests(TestCase):
 
         with self.assertRaises(FileNotFoundError):
             handle_weights_upload(path_to_model_to_upload=s3_path)
+
+    @skipIf(
+        not os.path.exists(os.path.join(os.path.expanduser("~"), ".cache", "deepparse", "cc.fr.300.bin")),
+        "download of model too long for test in runner",
+    )
+    def test_integration_handle_weights_with_uri(self):
+        uri = "s3://deepparse/fasttext.ckpt"
+
+        address_parser = AddressParser(model_type="fasttext", path_to_retrained_model=uri)
+        parse_address = address_parser("350 rue des Lilas Ouest Quebec city Quebec G1L 1B6")
+        self.assertIsInstance(parse_address, FormattedParsedAddress)
+
+        uri = CloudPath("s3://deepparse/fasttext.ckpt")
+
+        address_parser = AddressParser(model_type="fasttext", path_to_retrained_model=uri)
+        parse_address = address_parser("350 rue des Lilas Ouest Quebec city Quebec G1L 1B6")
+        self.assertIsInstance(parse_address, FormattedParsedAddress)
 
 
 if __name__ == "__main__":