+++ CHANGELOG.md
@@ -316,7 +316,7 @@
 ## dev
 
 - New models release with more meta-data
-- Add feature to use an AddressParser from a URI
-- Add feature to upload trained model to a URI
-- Add example on how to use URI for parsing from and to upload to
+- Add a feature to use an AddressParser from a URI
+- Add a feature to upload the trained model to a URI
+- Add an example of how to use URI for parsing from and uploading to
 - Improve error handling of `path_to_retrain_model`